Change datatype of distance from int16 to int32 to be able to calculate more than 65km

+++ firmware/components/sc_logic/sc_logic.c
@@ -53,7 +53,7 @@ static uint16_t calculate_distance() {
     float lat_km = lat_diff * 110.574;
     float lon_km = lon_diff * 111.320*cos(compass_data_ptr->path.nodes[display_data_ptr->next_wp].lat * M_PI / 180);
     
-    return (uint16_t)sqrt(lat_km * lat_km + lon_km * lon_km) * 1000;
+    return (uint32_t)sqrt(lat_km * lat_km + lon_km * lon_km) * 1000;
 }
 
 _Noreturn static void logic_task() {
